Added new console script to summarize a report.
Summary groups resources in a report by keystone domain then by
a user specified column and displays the sums of the costs at
each level.

Added a client to get keystone v3 domains from projects.
The openstack client was briefly tried but far too much of a hastle
to integrate. Instead, a keystoneclient adapter was used.

Added a new column field for project id. This existed previously
but was named payer_account_id.