Added a rough ability to gather licensing info.
The process is to run a special licensing report. The licensing
report should have columns looking for image metadata tags on
one of the meters that captures image metadata with instances.

Included in this change are a couple of licensers that can
examine the number of instances associated with software or that
can examine the amount of time instances with said software are in
use. These licensers can compute a cost according to some user
specified rates.